Revert the build script to how it used to be
