+++ package.json
@@ -43,15 +43,15 @@
     "typedoc-plugin-merge-modules": "^3.1.0",
     "typescript": "^4.7.4",
     "uuid": "^8.3.2",
-    "yargs": "^17.2.1",
-    "cmake-js": "^6.3.2",
-    "tar": "^6.1.11"
+    "yargs": "^17.2.1"
   },
   "dependencies": {
     "@aws-sdk/util-utf8-browser": "^3.109.0",
     "@httptoolkit/websocket-stream": "^6.0.0",
     "axios": "^0.24.0",
     "crypto-js": "^4.0.0",
-    "mqtt": "^4.3.7"
+    "mqtt": "^4.3.7",
+    "cmake-js": "^6.3.2",
+    "tar": "^6.1.11"
   }
 }
