--- scripts/build_dependencies/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "aws-crt-build",
-  "private": true,
-  "dependencies": {},
-  "devDependencies": {}
-}
